Added Parcel configuration
