+++ .cache/55/21fbc3ef3be3dd9fd4fc7f1a88d101.json
@@ -1,0 +1,1 @@
+{"id":"src/client/components/Header.js","dependencies":[{"name":"/Users/prateekgogia/Projects/grommerce-store/package.json","includedInParent":true,"mtime":1554265523929},{"name":"react","loc":{"line":1,"column":33},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/components/Header.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/index.js"}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.default = void 0;\n\nvar _react = _interopRequireWildcard(require(\"react\"));\n\nfunction _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj.default = obj; return newObj; } }\n\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nvar Header =\n/*#__PURE__*/\nfunction (_Component) {\n  _inherits(Header, _Component);\n\n  function Header() {\n    _classCallCheck(this, Header);\n\n    return _possibleConstructorReturn(this, _getPrototypeOf(Header).apply(this, arguments));\n  }\n\n  _createClass(Header, [{\n    key: \"render\",\n    value: function render() {\n      return _react.default.createElement(\"div\", null, \"Header\");\n    }\n  }]);\n\n  return Header;\n}(_react.Component);\n\nvar _default = Header;\nexports.default = _default;","map":{"mappings":[{"generated":{"line":8,"column":0},"source":"src/client/components/Header.js","original":{"line":1,"column":0}},{"name":"Header","generated":{"line":30,"column":4},"source":"src/client/components/Header.js","original":{"line":3,"column":6}},{"generated":{"line":43,"column":29},"source":"src/client/components/Header.js","original":{"line":4,"column":11}},{"generated":{"line":44,"column":0},"source":"src/client/components/Header.js","original":{"line":5,"column":4}},{"generated":{"line":44,"column":13},"source":"src/client/components/Header.js","original":{"line":5,"column":11}},{"generated":{"line":44,"column":64},"source":"src/client/components/Header.js","original":{"line":5,"column":4}},{"generated":{"line":45,"column":0},"source":"src/client/components/Header.js","original":{"line":6,"column":3}},{"name":"Component","generated":{"line":49,"column":2},"source":"src/client/components/Header.js","original":{"line":3,"column":21}},{"name":"Header","generated":{"line":51,"column":15},"source":"src/client/components/Header.js","original":{"line":9,"column":15}}],"sources":{"src/client/components/Header.js":"import React, { Component } from 'react';\n\nclass Header extends Component {\n  render() {\n    return <div>Header</div>;\n  }\n}\n\nexport default Header;\n"},"lineCount":null}},"hash":"ee574fd75e8fa068019ee567c2d656c9","cacheData":{"env":{}}}
