+++ .cache/c1/cb262860a73de476b4562a942693ec.json
@@ -1,0 +1,1 @@
+{"id":"index.html","dependencies":[{"name":"./src/client/index.js","dynamic":true,"resolved":"/Users/prateekgogia/Projects/grommerce-store/src/client/index.js","parent":"/Users/prateekgogia/Projects/grommerce-store/index.html"}],"generated":{"html":"<!DOCTYPE html>\n<html>\n  <head>\n    <title>React starter app</title>\n    <link rel=\"stylesheet\" href=\"https://cdnjs.cloudflare.com/ajax/libs/bulma/0.7.4/css/bulma.min.css\">\n  </head>\n  <body>\n    <div id=\"app\"></div>\n    <script src=\"/c4d47be494321bcb7989fbc811500f53.js\"></script>\n  </body>\n</html>\n"},"hash":"055dc44edfefa0eb43cf70095677dc19","cacheData":{}}
