+++ .cache/d1/3ba9cd0b7975914d1067a7055ff4ae.json
@@ -1,0 +1,1 @@
+{"id":"src/client/App.js","dependencies":[{"name":"/Users/prateekgogia/Projects/grommerce-store/package.json","includedInParent":true,"mtime":1554265523929},{"name":"react","loc":{"line":1,"column":33},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/App.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/index.js"},{"name":"./components/Header","loc":{"line":2,"column":19},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/App.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/src/client/components/Header.js"}],"generated":{"js":"\"use strict\";\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.default = void 0;\n\nvar _react = _interopRequireWildcard(require(\"react\"));\n\nvar _Header = _interopRequireDefault(require(\"./components/Header\"));\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj.default = obj; return newObj; } }\n\nfunction _typeof(obj) { if (typeof Symbol === \"function\" && typeof Symbol.iterator === \"symbol\") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === \"function\" && obj.constructor === Symbol && obj !== Symbol.prototype ? \"symbol\" : typeof obj; }; } return _typeof(obj); }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\nfunction _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === \"object\" || typeof call === \"function\")) { return call; } return _assertThisInitialized(self); }\n\nfunction _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return self; }\n\nfunction _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function\"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }\n\nfunction _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }\n\nvar App =\n/*#__PURE__*/\nfunction (_Component) {\n  _inherits(App, _Component);\n\n  function App() {\n    _classCallCheck(this, App);\n\n    return _possibleConstructorReturn(this, _getPrototypeOf(App).apply(this, arguments));\n  }\n\n  _createClass(App, [{\n    key: \"render\",\n    value: function render() {\n      return _react.default.createElement(\"div\", null, _react.default.createElement(_Header.default, null), _react.default.createElement(\"div\", {\n        className: \"container\"\n      }, _react.default.createElement(\"p\", null, \"Grommerce Store\")));\n    }\n  }]);\n\n  return App;\n}(_react.Component);\n\nvar _default = App;\nexports.default = _default;","map":{"mappings":[{"generated":{"line":8,"column":0},"source":"src/client/App.js","original":{"line":1,"column":0}},{"generated":{"line":10,"column":0},"source":"src/client/App.js","original":{"line":2,"column":0}},{"name":"App","generated":{"line":34,"column":4},"source":"src/client/App.js","original":{"line":4,"column":6}},{"generated":{"line":47,"column":29},"source":"src/client/App.js","original":{"line":5,"column":11}},{"generated":{"line":48,"column":0},"source":"src/client/App.js","original":{"line":6,"column":4}},{"generated":{"line":48,"column":13},"source":"src/client/App.js","original":{"line":7,"column":6}},{"generated":{"line":48,"column":55},"source":"src/client/App.js","original":{"line":8,"column":8}},{"generated":{"line":48,"column":84},"source":"src/client/App.js","original":{"line":8,"column":9}},{"generated":{"line":48,"column":99},"source":"src/client/App.js","original":{"line":8,"column":8}},{"generated":{"line":48,"column":106},"source":"src/client/App.js","original":{"line":7,"column":6}},{"generated":{"line":48,"column":108},"source":"src/client/App.js","original":{"line":9,"column":8}},{"generated":{"line":49,"column":0},"source":"src/client/App.js","original":{"line":9,"column":13}},{"generated":{"line":49,"column":8},"source":"src/client/App.js","original":{"line":9,"column":13}},{"generated":{"line":49,"column":17},"source":"src/client/App.js","original":{"line":9,"column":22}},{"generated":{"line":49,"column":19},"source":"src/client/App.js","original":{"line":9,"column":23}},{"generated":{"line":50,"column":0},"source":"src/client/App.js","original":{"line":9,"column":8}},{"generated":{"line":50,"column":9},"source":"src/client/App.js","original":{"line":10,"column":10}},{"generated":{"line":50,"column":67},"source":"src/client/App.js","original":{"line":9,"column":8}},{"generated":{"line":50,"column":68},"source":"src/client/App.js","original":{"line":7,"column":6}},{"generated":{"line":50,"column":69},"source":"src/client/App.js","original":{"line":6,"column":4}},{"generated":{"line":51,"column":0},"source":"src/client/App.js","original":{"line":14,"column":3}},{"name":"Component","generated":{"line":55,"column":2},"source":"src/client/App.js","original":{"line":4,"column":18}},{"name":"App","generated":{"line":57,"column":15},"source":"src/client/App.js","original":{"line":17,"column":15}}],"sources":{"src/client/App.js":"import React, { Component } from 'react';\nimport Header from './components/Header';\n\nclass App extends Component {\n  render() {\n    return (\n      <div>\n        <Header />\n        <div className=\"container\">\n          <p>Grommerce Store</p>\n        </div>\n      </div>\n    );\n  }\n}\n\nexport default App;\n"},"lineCount":null}},"hash":"682f0c67d98d2a439e31821c67cc83e2","cacheData":{"env":{}}}
