+++ .cache/e9/dbbbcb8736e470f18e2d5792c734fd.json
@@ -1,0 +1,1 @@
+{"id":"node_modules/react/index.js","dependencies":[{"name":"/Users/prateekgogia/Projects/grommerce-store/package.json","includedInParent":true,"mtime":1554265523929},{"name":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/package.json","includedInParent":true,"mtime":1554265541950},{"name":"./cjs/react.development.js","loc":{"line":6,"column":27},"parent":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/index.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/cjs/react.development.js"}],"generated":{"js":"'use strict';\n\nif (\"development\" === 'production') {\n  module.exports = require('./cjs/react.production.min.js');\n} else {\n  module.exports = require('./cjs/react.development.js');\n}","map":{"mappings":[{"generated":{"line":1,"column":0},"source":"node_modules/react/index.js","original":{"line":1,"column":0}},{"generated":{"line":3,"column":0},"source":"node_modules/react/index.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":4},"source":"node_modules/react/index.js","original":{"line":3,"column":4}},{"generated":{"line":3,"column":22},"source":"node_modules/react/index.js","original":{"line":3,"column":29}},{"generated":{"line":3,"column":34},"source":"node_modules/react/index.js","original":{"line":3,"column":0}},{"generated":{"line":3,"column":36},"source":"node_modules/react/index.js","original":{"line":3,"column":43}},{"name":"module","generated":{"line":4,"column":0},"source":"node_modules/react/index.js","original":{"line":4,"column":2}},{"name":"module","generated":{"line":4,"column":2},"source":"node_modules/react/index.js","original":{"line":4,"column":2}},{"generated":{"line":4,"column":8},"source":"node_modules/react/index.js","original":{"line":4,"column":8}},{"name":"exports","generated":{"line":4,"column":9},"source":"node_modules/react/index.js","original":{"line":4,"column":9}},{"generated":{"line":4,"column":16},"source":"node_modules/react/index.js","original":{"line":4,"column":2}},{"name":"require","generated":{"line":4,"column":19},"source":"node_modules/react/index.js","original":{"line":4,"column":19}},{"generated":{"line":4,"column":26},"source":"node_modules/react/index.js","original":{"line":4,"column":26}},{"generated":{"line":4,"column":27},"source":"node_modules/react/index.js","original":{"line":4,"column":27}},{"generated":{"line":4,"column":58},"source":"node_modules/react/index.js","original":{"line":4,"column":26}},{"generated":{"line":4,"column":59},"source":"node_modules/react/index.js","original":{"line":4,"column":2}},{"generated":{"line":5,"column":0},"source":"node_modules/react/index.js","original":{"line":5,"column":1}},{"generated":{"line":5,"column":1},"source":"node_modules/react/index.js","original":{"line":3,"column":0}},{"generated":{"line":5,"column":7},"source":"node_modules/react/index.js","original":{"line":5,"column":7}},{"name":"module","generated":{"line":6,"column":0},"source":"node_modules/react/index.js","original":{"line":6,"column":2}},{"name":"module","generated":{"line":6,"column":2},"source":"node_modules/react/index.js","original":{"line":6,"column":2}},{"generated":{"line":6,"column":8},"source":"node_modules/react/index.js","original":{"line":6,"column":8}},{"name":"exports","generated":{"line":6,"column":9},"source":"node_modules/react/index.js","original":{"line":6,"column":9}},{"generated":{"line":6,"column":16},"source":"node_modules/react/index.js","original":{"line":6,"column":2}},{"name":"require","generated":{"line":6,"column":19},"source":"node_modules/react/index.js","original":{"line":6,"column":19}},{"generated":{"line":6,"column":26},"source":"node_modules/react/index.js","original":{"line":6,"column":26}},{"generated":{"line":6,"column":27},"source":"node_modules/react/index.js","original":{"line":6,"column":27}},{"generated":{"line":6,"column":55},"source":"node_modules/react/index.js","original":{"line":6,"column":26}},{"generated":{"line":6,"column":56},"source":"node_modules/react/index.js","original":{"line":6,"column":2}},{"generated":{"line":7,"column":0},"source":"node_modules/react/index.js","original":{"line":7,"column":1}}],"sources":{"node_modules/react/index.js":"'use strict';\n\nif (process.env.NODE_ENV === 'production') {\n  module.exports = require('./cjs/react.production.min.js');\n} else {\n  module.exports = require('./cjs/react.development.js');\n}\n"},"lineCount":null}},"hash":"98fe26513a4bd17df292b8509cd3dd3f","cacheData":{"env":{"NODE_ENV":"development"}}}
