+++ .cache/eb/dd3fd81fb2414262fa2bf876cc9916.json
@@ -1,0 +1,1 @@
+{"id":"src/client/index.js","dependencies":[{"name":"/Users/prateekgogia/Projects/grommerce-store/package.json","includedInParent":true,"mtime":1554265523929},{"name":"react","loc":{"line":1,"column":18},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/index.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react/index.js"},{"name":"react-dom","loc":{"line":2,"column":21},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/index.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/node_modules/react-dom/index.js"},{"name":"./App","loc":{"line":3,"column":16},"parent":"/Users/prateekgogia/Projects/grommerce-store/src/client/index.js","resolved":"/Users/prateekgogia/Projects/grommerce-store/src/client/App.js"}],"generated":{"js":"\"use strict\";\n\nvar _react = _interopRequireDefault(require(\"react\"));\n\nvar _reactDom = _interopRequireDefault(require(\"react-dom\"));\n\nvar _App = _interopRequireDefault(require(\"./App\"));\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\n// import './scss/app.scss';\n_reactDom.default.render(_react.default.createElement(_App.default, null), document.getElementById('app'));","map":{"mappings":[{"generated":{"line":3,"column":0},"source":"src/client/index.js","original":{"line":1,"column":0}},{"generated":{"line":5,"column":0},"source":"src/client/index.js","original":{"line":2,"column":0}},{"generated":{"line":7,"column":0},"source":"src/client/index.js","original":{"line":3,"column":0}},{"generated":{"line":11,"column":0},"source":"src/client/index.js","original":{"line":4,"column":0}},{"name":"ReactDOM","generated":{"line":12,"column":0},"source":"src/client/index.js","original":{"line":6,"column":0}},{"name":"render","generated":{"line":12,"column":18},"source":"src/client/index.js","original":{"line":6,"column":9}},{"generated":{"line":12,"column":24},"source":"src/client/index.js","original":{"line":6,"column":0}},{"generated":{"line":12,"column":25},"source":"src/client/index.js","original":{"line":6,"column":16}},{"generated":{"line":12,"column":54},"source":"src/client/index.js","original":{"line":6,"column":17}},{"generated":{"line":12,"column":66},"source":"src/client/index.js","original":{"line":6,"column":16}},{"generated":{"line":12,"column":73},"source":"src/client/index.js","original":{"line":6,"column":0}},{"name":"document","generated":{"line":12,"column":75},"source":"src/client/index.js","original":{"line":6,"column":25}},{"generated":{"line":12,"column":83},"source":"src/client/index.js","original":{"line":6,"column":33}},{"name":"getElementById","generated":{"line":12,"column":84},"source":"src/client/index.js","original":{"line":6,"column":34}},{"generated":{"line":12,"column":98},"source":"src/client/index.js","original":{"line":6,"column":25}},{"generated":{"line":12,"column":99},"source":"src/client/index.js","original":{"line":6,"column":49}},{"generated":{"line":12,"column":104},"source":"src/client/index.js","original":{"line":6,"column":25}},{"generated":{"line":12,"column":105},"source":"src/client/index.js","original":{"line":6,"column":0}}],"sources":{"src/client/index.js":"import React from 'react';\nimport ReactDOM from 'react-dom';\nimport App from './App';\n// import './scss/app.scss';\n\nReactDOM.render(<App />, document.getElementById('app'));\n"},"lineCount":null}},"hash":"148af8dbca3351a7c8b8b8d99bd0f655","cacheData":{"env":{}}}
