+++ index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>React starter app</title>
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/bulma/0.7.4/css/bulma.min.css"
+    />
+  </head>
+  <body>
+    <div id="app"></div>
+    <script src="./src/client/index.js"></script>
+  </body>
+</html>
