+++ src/client/App.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import Header from './components/Header';
+
+class App extends Component {
+  render() {
+    return (
+      <div>
+        <Header />
+        <div className="container">
+          <p>Grommerce Store</p>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default App;
