+++ src/client/components/Header.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+class Header extends Component {
+  render() {
+    return <div>Header</div>;
+  }
+}
+
+export default Header;
