+++ src/client/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
+// import './scss/app.scss';
+
+ReactDOM.render(<App />, document.getElementById('app'));
