+++ src/client/pages/index.js
@@ -1,0 +1,12 @@
+import Link from 'next/link';
+
+const Index = () => (
+  <div>
+    <Link href="/about">
+      <a>About Page</a>
+    </Link>
+    <p>Hello Next.js</p>
+  </div>
+);
+
+export default Index;
